Move interop to bottom of import
