+++ cirq/interop/qasm/qasm_input.py
@@ -11,3 +11,19 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from cirq import circuits
+from cirq.interop.qasm._parser import QasmParser
+
+
+def circuit_from_qasm(qasm: str) -> circuits.Circuit:
+    """Parses an OpenQASM string to `cirq.Circuit`.
+
+    Args:
+        qasm: The OpenQASM string
+
+    Returns:
+        The parsed circuit
+    """
+
+    return QasmParser().parse(qasm).circuit
